Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 # include <stdio.h>
 
 int main(){
-    int num;
+    long double num;
     scanf("Lf",&num);
     printf("You entered: %.4Lf",num);
     return 0;
